Modals finished. Need to xfer all to React
