+++ js/modules/signal.js
@@ -50,6 +50,7 @@ const {
 } = require('../../ts/components/conversation/CreateGroupDialog');
 const { EditProfileDialog } = require('../../ts/components/EditProfileDialog');
 const { UserDetailsDialog } = require('../../ts/components/UserDetailsDialog');
+const { DevicePairingDialog } = require('../../ts/components/DevicePairingDialog');
 const { SessionToast } = require('../../ts/components/session/SessionToast');
 const { SessionToggle } = require('../../ts/components/session/SessionToggle');
 const { SessionModal } = require('../../ts/components/session/SessionModal');
@@ -248,6 +249,7 @@ exports.setup = (options = {}) => {
     CreateGroupDialog,
     EditProfileDialog,
     UserDetailsDialog,
+    DevicePairingDialog,
     SessionRegistrationView,
     ConfirmDialog,
     UpdateGroupDialog,
